Raster: Fix issue where rasters appear behind tiles by default
The default zIndex for TileLayer and GridLayer is now 1. In
Leaflet 0.7 it was null. In our R code, the default was NULL,
which we pass to the Leaflet code. Therefore, in Leaflet 1.x
we actually override the default zIndex of 1 and change it to
"auto".

However, for addRaster, we've never allowed the zIndex to be
set, so we don't override the new default zIndex of 1. So we
end up with a tile layer with zIndex "auto" and a grid layer
with a zIndex of 1. For "stacking context" reasons that I do
not understand, auto always renders in front of non-auto.

Setting the default for tiles to 1 fixes the problem.