Fix issue with rasters disappearing
The getImageData(callback) function as previously written
would invoke the callback immediately if image data was
already loaded, or asynchronously if image data was not.
(I knew this was weird so I noted this in the function's
comment.) This turns out to be a problem with Leaflet 1.x
because it expects createTile's callback (done) function
to only be invoked after createTile has returned. If done
is called too soon, then it's ignored, because done first
checks the GridLayer's tile cache to see if the tile that's
done is even still wanted. Since createTile hasn't returned
the tile isn't in the cache yet, and done becomes a no-op.